+++ src/test/groovy/nextflow/slack/SlackConfigTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025, Seqera Labs
+ * Copyright 2025-2026, Seqera Labs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -496,5 +496,33 @@ class SlackConfigTest extends Specification {
         then:
         config.seqeraPlatform != null
         config.seqeraPlatform.enabled == false
+    }
+
+    def 'should parse seqeraPlatform actionButtons config'() {
+        given:
+        def session = Mock(Session) {
+            config >> [slack: [
+                webhook: [url: 'https://hooks.slack.com/test'],
+                seqeraPlatform: [
+                    enabled: true,
+                    actionButtons: [
+                        mode: 'interactive',
+                        cancel: false,
+                        resume: true,
+                        relaunch: false
+                    ]
+                ]
+            ]]
+        }
+
+        when:
+        def config = SlackConfig.from(session)
+
+        then:
+        config.seqeraPlatform.actionButtons.mode == 'interactive'
+        !config.seqeraPlatform.actionButtons.cancel
+        config.seqeraPlatform.actionButtons.resume
+        !config.seqeraPlatform.actionButtons.relaunch
+        config.seqeraPlatform.actionButtons.isInteractiveMode()
     }
 }